refactor: consolidate type definitions
