+++ apps/campfire/src/components/Deck/Slide/context.ts
@@ -1,5 +1,5 @@
 import { createContext } from 'preact'
-import type { Transition } from './types'
+import type { Transition } from '@campfire/types'
 
 /**
  * Context providing default transitions for SlideReveal components within a Slide.
